+++ Craftsman/Models/MessageTemplate.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Craftsman.Models
+{
+    public class MessageTemplate
+    {
+        public List<Message> Messages { get; set; }
+    }
+}